Rename server/deps.py to server/query_params.py
