+++ optimade/server/routers/structures.py
@@ -10,9 +10,9 @@ from optimade.models import (
     StructureResponseOne,
 )
 from optimade.server.config import CONFIG
-from optimade.server.deps import EntryListingQueryParams, SingleEntryQueryParams
 from optimade.server.entry_collections import MongoCollection, client
 from optimade.server.mappers import StructureMapper
+from optimade.server.query_params import EntryListingQueryParams, SingleEntryQueryParams
 
 from .utils import get_entries, get_single_entry
